pyproject.toml
[tool.poetry]
name = "mirai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
graia-ariadne = "^0.6.11"
numpy = "1.19.3"
Pillow = "^9.0.1"
saucenao-api = "^2.4.0"
arclet-alconna-graia = "^0.0.4"
httpx = "^0.22.0"
matplotlib = "^3.5.1"
PyMySQL = "^1.0.2"
graia-saya = "^0.0.15"
graia-scheduler = "^0.0.7"
playwright = "^1.20.1"
paddleocr = "^2.4.0"
paddlepaddle = "^2.2.2"
graiax-silkcoder = "0.1.1"
ffmpeg = "^1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
